LIBITD-2130. Updated "wstrack" client to use Java 8 and other fixes
Made the following updates to the repository:

1) Delete the "Dockerfile", as it no longer works now that the "server"
is using Rails.

2) Removed the root "pom.xml" as it is no longer needed (the
"client/pom.xml" is the only POM file that is necessary).

3) Updated the "client/pom.xml" file with the necessary configuration
from the delete root "pom.xml" file. Also add Java 1.7 as the
"compilation target", to ensure that the resulting jar could be used
with a Java 7 JRE.

4) Updated the client/README.md

5) Fixed compilation errors in "ClientTracking.java", related to the
"feature/LIBITD-396" rebase on the "main" branch.

https://issues.umd.edu/browse/LIBITD-2130