LIBITD-2136. Updated "config.hosts" regex for IP addresses
Fixed an issue where the readiness/liveness probes in the "wstrack"
Kubernetes stack received an HTTP 403 (Forbidden) error when attempting
to access the "/ping" endpoint.

The initial "config.hosts" configuration was taken from the "umd-handle"
application, which is on Rails v6.1.3.1. In Rails v6.1.4.4, Rails
does additional processing of the regex to enable a port number of be
included. Having the "\A...\z" directives on the IP address regex
was interfering with the processing, and so was removed in this change.

Note: Rails itself wraps the regex in "\A...\z" directives, so they
are redundant, in any case.

https://issues.umd.edu/browse/LIBITD-2136