LIBITD-2117. Modified History recording to be handled by controller
Having the history recorded using an "after_commit" hook for new records
was not working as expected, because the wstrack client endpoint will
look up and modify an existing workstation status record.

Since we are only interested in recording login/logout activity on the
wstrack client endpoint, placed the history recording in the
controller.

Updated the tests.

https://issues.umd.edu/browse/LIBITD-2117