LIBITD-2126. Ensure Grails times are parsed in proper timezone
Modified the "import_current_workstation_status" Rails task in
"server/lib/tasks/grails_migration.rake" to use the application
time zone when parsing the timestamps in the imported CSV file.

This is to ensure that the timestamps are treated being in
the EST/EDT timezone, instead of the UTC timezone.

https://issues.umd.edu/browse/LIBITD-2126