LIBITD-2142. Removed "id" field from CSV history file
Removed the "id" field from the CSV history file, as it is not
particularly useful, as there will be multiple instances of the same
"id" in a CSV file, as WorkstationStatus entries that already exist are
updated, not created for each client access.
So when a particular workstation reports a login and later reports a
logout, the CSV file will contain two entries with the same "id" – the
id of the WorkstationStatus record in the database.

This seemed confusing as it was not in keeping with the usual definition
of an id field. Additionally, the id field contains no particularly
useful information.

https://issues.umd.edu/browse/LIBITD-2142