LIBITD-2186. Modified Grouper group comparison to be case-insensitive
Modified the "roles_allowed?" method in SessionsController to do
a case-insensitive comparison to the Grouper Group, as it may be
returned in different uppercase/lowercase forms.

Added a test case for case insensitivity, and modified the
"login_with_valid_role" method in test_helper.rb to take an
optional role argument.

https://issues.umd.edu/browse/LIBITD-2186